+++ modules/video_filter/rotate.c
@@ -92,11 +92,11 @@ struct filter_sys_t
 
 static void store_trigo( struct filter_sys_t *sys, float f_angle )
 {
-    f_angle *= M_PI / 180.f; /* degrees -> radians */
+    f_angle *= (float)(M_PI / 180.); /* degrees -> radians */
 
     uint16_t i_sin = lroundf(sinf(f_angle) * 4096.f);
     uint16_t i_cos = lroundf(cosf(f_angle) * 4096.f);
-    atomic_store( &sys->sincos, (i_cos << 16) | (i_sin << 0));
+    atomic_store( &sys->sincos, (i_cos << 16u) | (i_sin << 0u));
 }
 
 static void fetch_trigo( struct filter_sys_t *sys, int *i_sin, int *i_cos )
